временная свалка в app
