+++ .gitignore
@@ -25,3 +25,4 @@ venv/
 venv
 
 .env
+/postgres-data/
