--- users/Login.py
@@ -1,21 +1,0 @@
-
-class UserLogin():
-    def fromDB(self, user_id, db):
-        self.__user = db.getUser(user_id)
-        return self
-
-    def create(self, user):
-        self.__user = user
-        return self
-
-    def is_authentificated(self):
-        return True
-
-    def is_active(self):
-        return True
-
-    def is_anonymus(self):
-        return False
-
-    def get_id(self):
-        return str(self.user['id'])
